1. Fixed the double HTML entity escaping problem with chapter headings 2. Added some logging in as any parent styles are missing
